feat: move getTradeQuote to getMultiHopTradeQuote
